separate test from global_muscle_tracking
